Refactor language switching and remove i18n dependencies
The commit includes changes to the language switching functionality in the Switchlanguage component, removing dependencies on the i18n library. The RootLayout component also had some changes made to it, including the removal of i18n imports and the addition of metadata. The test component had some minor changes made to it as well.