Add Riyad Bank font and update language support
This commit adds the Riyad Bank font and updates the CSS to use it where appropriate. It also updates the language support to include Arabic and adds translations for the "freelancers" text. The TranslateContext component has been updated to set the direction of the page based on the selected language and the font family has been set to Riyad Bank.